MOVE: CommonRequest, CommonResponse to slatekit.common.requests
