+++ src/lib/kotlin/slatekit-apis/src/main/kotlin/slatekit/apis/core/Requests.kt
@@ -21,6 +21,7 @@ import slatekit.common.*
 import slatekit.common.encrypt.Encryptor
 import slatekit.common.ext.toStringUtc
 import slatekit.common.io.Uris
+import slatekit.common.requests.CommonRequest
 import slatekit.common.requests.Request
 import slatekit.meta.Serialization
 
@@ -93,16 +94,16 @@ object Requests {
         val sep = if (path.contains("/")) "/" else "."
 
         return CommonRequest(
-                version = version,
-                path = path,
-                parts = path.split(sep),
-                source = Source.parse(sourceOverride ?: source),
-                verb = verbOverride ?: verb,
-                meta = Meta(rawSource ?: "json", jsonMeta, enc),
-                data = Data(rawSource ?: "json", Source.File.id, true, enc, jsonData),
-                raw = rawSource,
-                tag = tag,
-                timestamp = time
+            version = version,
+            path = path,
+            parts = path.split(sep),
+            source = Source.parse(sourceOverride ?: source),
+            verb = verbOverride ?: verb,
+            meta = Meta(rawSource ?: "json", jsonMeta, enc),
+            data = Data(rawSource ?: "json", Source.File.id, true, enc, jsonData),
+            raw = rawSource,
+            tag = tag,
+            timestamp = time
         )
     }
