+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/apis/Api_Security_Tests.kt
@@ -17,7 +17,7 @@ import slatekit.apis.core.Api
 import slatekit.apis.SetupType
 import slatekit.apis.Verbs
 import slatekit.common.info.Credentials
-import slatekit.common.CommonRequest
+import slatekit.common.requests.CommonRequest
 import slatekit.common.Source
 import slatekit.common.ext.toResponse
 import slatekit.results.Success
