+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/meta/CallTests.kt
@@ -17,7 +17,7 @@ import org.junit.Test
 import slatekit.common.Conversions
 import slatekit.common.DateTimes
 import slatekit.common.requests.InputArgs
-import slatekit.common.CommonRequest
+import slatekit.common.requests.CommonRequest
 import slatekit.common.Source
 import slatekit.meta.Reflector
 import slatekit.meta.Deserializer
